Finalizing the light puzzle
-New "Prism" script.
Objects with the "Prism" script can now have a fixed target on which they always direct the light beam.
Objects with the "Prism" script can now activate other objects with the "Prism" script, provided that this other object is not already activated.

With both the Lightstone and the prism, you can now set whether the light beam should be dynamically adjusted to the alignment and position of the object that is used as the source of the light beam during runtime.

If you deactivate a Lightstone or a prism or repositioning it or change its rotation, all prisms that were dependent on this light source are also deactivated