Finalization of the Orb's Path Finding method
You can now give the orb any number of "stages".
Each "Stage" object has any number of "PathContainer" objects, which in turn all contain exactly 4 waypoints.
Each "PathContainer" represents a Bézier curve. Thus every "stage" is a series of Bézier curves.

That means it is now possible to specify any number of "stages" (path courses) and let the orb follow them in the specified order.

In the "OrbStageTransition.cs", an intermediate step can be defined for each transition from one "Stage" to the next.
The transition is automatically recognized (e.g. from "Stage 1" to "Stage 2"), but if there are more than 4 "Stages", new functions must be created in the "OrbStageTransition.cs", as this does not happen dynamically.