Better player orientation
Creation of HideTreesInRange.cs:
 Objects that have a collider and the tag "ObstacleTree"
 now disappear if they are inside the "TriggerHolder" trigger and their world position has a lower Z-value than the player

In addition, the LOD properties were removed from all trees in the Obstacle Prefabs (these are not necessary anyway) and the trees correspondingly provided with colliders that are only used for the HideTreesInRange.cs. In addition, the "ObstacleTree" tag was assigned to each of these trees.