better ground detection
the player is now not immediately marked as "not grounded" if there is no floor under him, but only after 0.1 seconds. If the ground contact should be restored during this time, the process of "marking as not grounded" will be canceled. in this way, small bumps in the terrain are taken into account by the ground detection and the player no longer glides over the ground in some places as before